Rename `ClientBaseOptimizedParts.{h, cpp}` to `ClientApplicationBaseParser.{h, cpp}`
The files only include the parsing logic for `ClientApplicationBase`.
